Git Gud v0.2.9 GitGud.CommentQuery

Conveniences for comment related queries.

Functions

by_id(id, opts \\ [])

by_id(pos_integer(), keyword()) :: GitGud.Comment.t() | nil

Returns a comment for the given id.

comment_query(id)

comment_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a single comment by id.

thread(comment, opts \\ [])

Returns the thread associated to the given comment.

thread_query(id, table)

thread_query(pos_integer(), binary()) :: Ecto.Query.t()

Returns a query for fetching the associated thread for the given comment id and table.