Git Gud v0.2.9 GitGud.CommitQuery

Conveniences for commit related queries.

Functions

ancestors_count_query(repo, commit)

ancestors_count_query(
  GitGud.Repo.t() | pos_integer(),
  GitGud.Commit.t() | Git.oid()
) :: Ecto.Query

Returns a query for counting the number of ancestors for the given commit.

by_oid(repo, oid, opts \\ [])

by_oid(GitGud.Repo.t() | pos_integer(), Git.oid(), keyword()) ::
  GitGud.Commit.t() | nil

Returns a commit for the given oid.

commit_query(repo_id, oid)

commit_query(GitGud.Repo.t() | pos_integer(), Git.oid()) :: Ecto.Query

Returns a query for fetching the commit with the given oid.

count_ancestors(repo, commit, opts \\ [])

count_ancestors(
  GitGud.Repo.t() | pos_integer(),
  GitGud.Commit.t() | Git.oid(),
  keyword()
) :: non_neg_integer()

Returns the number of ancestors for the given repo and commit.

gpg_signature(repo, commit, opts \\ [])

gpg_signature(
  GitGud.Repo.t() | pos_integer(),
  GitGud.Commit.t() | Git.oid() | [GitGud.Commit.t() | Git.oid()],
  keyword()
) :: GitGud.GPGKey.t() | map()

Returns the GPG key associated to given signed commit.

gpg_signature_query(repo, commit)

gpg_signature_query(
  GitGud.Repo.t() | pos_integer(),
  GitGud.Commit.t() | Git.oid() | [GitGud.Commit.t() | Git.oid()]
) :: Ecto.Query

Returns a query for fetching the GPG key associated to the given signed commit.

history(repo, commit, opts \\ [])

history(
  GitGud.Repo.t() | pos_integer(),
  GitGud.Commit.t() | Git.oid(),
  keyword()
) :: [GitGud.Commit.t()]

Returns the commit history starting from the given commit.

history_query(repo, commit)

history_query(GitGud.Repo.t() | pos_integer(), GitGud.Commit.t() | Git.oid()) ::
  Ecto.Query

Returns a query for fetching the commit history starting from the given commit.