Git Gud v0.2.9 GitGud.GraphQL.Resolvers

Module providing resolution functions for GraphQL related queries.

Functions

close_issue(parent, args, info)

close_issue(any(), %{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Issue.t()} | {:error, term()}

Closes a repository issue.

comment_author(comment, args, info)

comment_author(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the author for a given comment.

comment_deletable(comment, args, info)

comment_deletable(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, boolean()} | {:error, term()}

Returns true if the viewer can delete a given comment; otherwise, returns false.

comment_deleted(args, info)

comment_deleted(map(), map()) :: {:ok, keyword()} | {:error, term()}

Returns the subscription topic for comment delete events.

comment_editable(comment, args, info)

comment_editable(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, boolean()} | {:error, term()}

Returns true if the viewer can edit a given comment; otherwise, returns false.

comment_html(comment, args, info)

comment_html(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves the HTML content of a given comment.

comment_repo(comment, args, info)

comment_repo(GitGud.Comment.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves the repository for a given comment.

comment_updated(args, info)

comment_updated(map(), map()) :: {:ok, keyword()} | {:error, term()}

Returns the subscription topic for comment update events.

commit_line_review(commit, args, info)

commit_line_review(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.CommitLineReview.t()} | {:error, term()}

Resolves the line review for a given Git commit object.

commit_line_review_comment_topic(map, info)

commit_line_review_comment_topic(map(), map()) ::
  {:ok, keyword()} | {:error, term()}

Returns the subscription topic for commit line review comment events.

commit_line_review_comments(args, info)

commit_line_review_comments(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves comments for a commit line review.

commit_line_review_created(map, info)

commit_line_review_created(map(), map()) ::
  {:ok, keyword()} | {:error, term()}

Returns the subscription topic for commit line review create event.

commit_line_reviews(commit, args, info)

commit_line_reviews(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [GitGud.CommitLineReview.t()]} | {:error, term()}

Resolves the line reviews for a given Git commit object.

commit_review(commit, args, info)

commit_review(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.CommitLineReview.t()} | {:error, term()}

Resolves the review for a given Git commit object.

commit_review_comment_topic(map, info)

commit_review_comment_topic(map(), map()) ::
  {:ok, keyword()} | {:error, term()}

Returns the subscription topic for commit review comment events.

commit_review_comments(args, info)

commit_review_comments(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves comments for a commit review.

create_commit_line_review_comment(parent, args, info)

create_commit_line_review_comment(
  any(),
  %{
    repo_id: pos_integer(),
    commit_oid: Git.oid(),
    blob_oid: Git.oid(),
    hunk: non_neg_integer(),
    line: non_neg_integer(),
    body: binary()
  },
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Creates a Git line commit review comment.

create_commit_review_comment(parent, args, info)

create_commit_review_comment(
  any(),
  %{repo_id: pos_integer(), commit_oid: Git.oid(), body: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Creates a Git commit review commit.

create_issue_comment(parent, args, info)

create_issue_comment(
  any(),
  %{id: pos_integer(), body: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Creates a repository issue comment.

delete_comment(parent, args, info)

delete_comment(any(), %{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Comment.t()} | {:error, term()}

Deletes a comment.

git_actor_type(actor, info)

git_actor_type(GitGud.User.t() | map(), Absinthe.Resolution.t()) :: atom()

Resolves the type for a given Git actor.

git_blob_size(blob, args, info)

git_blob_size(GitRekt.GitBlob.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves the content length for a given Git blob object.

git_commit_author(commit, args, info)

git_commit_author(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git commit object.

git_commit_committer(commit, args, info)

git_commit_committer(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the committer for a given Git commit object.

git_commit_message(commit, args, info)

git_commit_message(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git commit object.

git_commit_parents(args, info)

git_commit_parents(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the parents for a given Git commit object.

git_commit_timestamp(commit, args, info)

git_commit_timestamp(GitRekt.GitCommit.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, DateTime.t()} | {:error, term()}

Resolves the timestamp for a given Git commit object.

git_history(args, info)

git_history(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the commit history starting from the given Git revision object.

git_object_type(object, info)

git_object_type(GitRekt.GitAgent.git_object(), Absinthe.Resolution.t()) ::
  atom()

Resolves the type for a given Git object.

git_reference_target(reference, args, info)

git_reference_target(GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves the Git target for the given Git reference object.

git_reference_type(reference, args, info)

git_reference_type(GitReference.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, atom()} | {:error, term()}

Resolves the type for a given Git tag.

git_tag_author(tag, args, info)

git_tag_author(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t() | map()} | {:error, term()}

Resolves the author for a given Git tag object.

git_tag_message(tag, args, info)

git_tag_message(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()} | {:error, term()}

Resolves the message for a given Git tag object.

git_tag_target(tag, args, info)

git_tag_target(GitRekt.GitTag.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves the Git target for the given Git tag object.

git_tag_type(ref, info)

git_tag_type(GitReference.t() | GitRekt.GitTag.t(), Absinthe.Resolution.t()) ::
  {:ok, atom()} | {:error, term()}

Resolves the type for a given Git tag.

git_tree(revision, args, info)

git_tree(GitGud.Repo.git_revision(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitTree.t()} | {:error, term()}

Resolves the tree for a given Git commit object.

git_tree_entries(args, info)

git_tree_entries(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the tree entries for a given Git tree object.

git_tree_entries_with_last_commit(args, info)

git_tree_entries_with_last_commit(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the tree entries and their associated commit for a given pathspec.

git_tree_entry_target(tree_entry, args, info)

git_tree_entry_target(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitTree.t() | GitRekt.GitBlob.t()} | {:error, term()}

Returns the underlying Git object for a given Git tree_entry object.

git_tree_entry_with_last_commit(map, info)

git_tree_entry_with_last_commit(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves the tree entries and their associated commit for a given pathspec.

issue_author(issue, args, info)

issue_author(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the author for a given comment.

issue_comments(args, info)

issue_comments(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves comments for an issue.

issue_commit_reference_event_oid(event, args, info)

issue_commit_reference_event_oid(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()}

Resolves the commit OID for a given issue event.

issue_commit_reference_event_url(map, args, info)

issue_commit_reference_event_url(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, binary()}

Resolves the commit URL for a given issue event.

issue_editable(issue, args, info)

issue_editable(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, boolean()} | {:error, term()}

Returns true if the viewer can edit a given issue; otherwise, returns false.

issue_event_field(event, field, args, info)

issue_event_field(map(), binary(), %{}, Absinthe.Resolution.t()) ::
  {:ok, any()}

Resolves the field of a given issue event.

issue_event_timestamp(event, args, info)

issue_event_timestamp(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, NaiveDateTime.t()} | {:error, term()}

Resolves the timestamp for a given issue event.

issue_event_type(event, info)

issue_event_type(map(), Absinthe.Resolution.t()) :: atom()

Resolves the type of a given issue event.

issue_event_user(event, args, info)

issue_event_user(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves the user for a given issue event.

issue_labels_update_event_pull_labels(event, args, info)

issue_labels_update_event_pull_labels(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [binary()]}

Resolves the pull labels ids for a given issue event.

issue_labels_update_event_push_labels(event, args, info)

issue_labels_update_event_push_labels(map(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [binary()]}

Resolves the push labels ids for a given issue event.

issue_repo(issue, args, info)

issue_repo(GitGud.Issue.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves the repository for a given issue.

issue_topic(map, info)

issue_topic(map(), map()) :: {:ok, keyword()} | {:error, term()}

Returns the subscription topic for issue comment events.

node(node_type, info, opts \\ [])

node(map(), Absinthe.Resolution.t(), keyword()) ::
  {:ok, map()} | {:error, term()}

Resolves a node object.

node_type(node, info)

node_type(map(), Absinthe.Resolution.t()) :: atom() | nil

Resolves a node object type.

reopen_issue(parent, args, info)

reopen_issue(any(), %{id: pos_integer()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Issue.t()} | {:error, term()}

Reopens a repository issue.

repo(args, info)

repo(map(), Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves a repository object.

repo_head(repo, args, info)

repo_head(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t()} | {:error, term()}

Resolves the default branch object for a given repo.

repo_issue(repo, args, info)

repo_issue(GitGud.Repo.t(), %{number: number()}, Absinthe.Resolution.t()) ::
  {:ok, integer()} | {:error, term()}

Resolves a single issue of a given repo.

repo_issue_labels(repo, args, info)

repo_issue_labels(GitGud.Repo.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, [GitGud.IssueLabel.t()]}

Resolvers the issue labels associated to a given repo.

repo_issues(args, info)

repo_issues(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves issues for a repository.

repo_object(repo, args, info)

repo_object(GitGud.Repo.t(), %{oid: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves a Git object by OID for a given repo.

repo_ref(repo, args, info)

repo_ref(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t()} | {:error, term()}

Resolves a Git reference object by name for a given repo.

repo_refs(args, info)

repo_refs(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git reference objects for a given repo.

repo_revision(repo, args, info)

repo_revision(GitGud.Repo.t(), %{spec: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitRekt.GitAgent.git_object()} | {:error, term()}

Resolves a Git commit by revision spec for a given repo.

repo_tag(repo, args, info)

repo_tag(GitGud.Repo.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t() | GitRekt.GitTag.t()} | {:error, term()}

Resolves a Git tag object by name for a given repo.

repo_tags(args, info)

repo_tags(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all Git tag objects for a given repo.

search(args, info)

search(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves a list of users for a given search term.

search_result_type(result, info)

search_result_type(GitGud.User.t() | GitGud.Repo.t(), Absinthe.Resolution.t()) ::
  atom()

Returns the search result type.

update_comment(parent, args, info)

update_comment(
  any(),
  %{id: pos_integer(), body: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Comment.t()} | {:error, term()}

Updates a comment.

update_issue_labels(parent, args, info)

update_issue_labels(
  any(),
  %{id: pos_integer(), label_id: pos_integer()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Issue.t()} | {:error, term()}

Updates the label of an issue.

update_issue_title(parent, args, info)

update_issue_title(
  any(),
  %{id: pos_integer(), title: binary()},
  Absinthe.Resolution.t()
) :: {:ok, GitGud.Issue.t()} | {:error, term()}

Updates the title of an issue.

url(resource, args, info)

url(map(), %{}, Absinthe.Resolution.t()) :: {:ok, binary()} | {:error, term()}

Resolves the URL of the given resource.

user(root, args, info)

user(%{}, %{login: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.User.t()} | {:error, term()}

Resolves an user object by login.

user_public_email(user, args, info)

user_public_email(GitGud.User.t(), %{}, Absinthe.Resolution.t()) ::
  {:ok, GitReference.t()} | {:error, term()}

Resolves the public email for a given user.

user_repo(user, args, info)

user_repo(GitGud.User.t(), %{name: binary()}, Absinthe.Resolution.t()) ::
  {:ok, GitGud.Repo.t()} | {:error, term()}

Resolves a repository object by name for a given user.

user_repos(args, info)

user_repos(map(), Absinthe.Resolution.t()) ::
  {:ok, Absinthe.Relay.Connection.t()} | {:error, term()}

Resolves all repositories for an user.