Git Gud v0.2.9 GitGud.IssueQuery

Conveniences for issue related queries.

Functions

by_id(id, opts \\ [])

by_id(pos_integer(), keyword()) :: GitGud.Issue.t() | nil

Returns a repository issue for the given id.

comments_query(issue)

comments_query(GitGud.Issue.t() | pos_integer()) :: Ecto.Query.t()

Returns a query for fetching comments of an issue.

count_repo_issues(repo, opts \\ [])

count_repo_issues(GitGud.Repo.t() | pos_integer(), keyword()) ::
  non_neg_integer() | nil

Returns the number of issues for the given repo.

count_repo_issues_query(repo_id, status \\ :all)

count_repo_issues_query(pos_integer(), atom()) :: Ecto.Query.t()

Returns a query for counting the number of issues of a repository.

issue_query(id)

issue_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a repository issue by its id.

repo_issue(repo, number, opts \\ [])

repo_issue(GitGud.Repo.t() | pos_integer(), pos_integer(), keyword()) ::
  GitGud.Issue.t() | nil

Returns a repository issue for the given number.

repo_issue_query(repo_id, number)

repo_issue_query(pos_integer(), pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a repository issue by its number.

repo_issues(repo, opts \\ [])

repo_issues(GitGud.Repo.t() | pos_integer(), keyword()) :: [GitGud.Issue.t()]

Returns all issues for the given repo.

repo_issues_query(repo_id)

repo_issues_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching all repository issues.

repo_issues_query(repo_id, numbers_or_status)

repo_issues_query(pos_integer(), [pos_integer()] | atom()) :: Ecto.Query.t()

Returns a query for fetching all repository issues with the given status.

repo_issues_with_comments_count(repo, opts \\ [])

repo_issues_with_comments_count(GitGud.Repo.t() | pos_integer(), keyword()) :: [
  {GitGud.Issue.t(), pos_integer()}
]

Returns all issues and their number of comments for the given repo.

repo_issues_with_comments_count_query(repo_id, status)

repo_issues_with_comments_count_query(pos_integer(), atom()) :: Ecto.Query.t()

Returns a query for fetching all repository issues and their number of comments.

repo_labels(repo, opts \\ [])

repo_labels(GitGud.Repo.t() | pos_integer(), keyword()) :: [
  GitGud.IssueLabel.t()
]

Returns all issue labels for the given repo.

repo_labels_query(repo_id)

repo_labels_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching all issue labels of a repository.