Git Gud v0.2.9 GitGud.Mailer

Conveniences for composing and sending emails.

Functions

deliver(email)

deliver(any()) :: no_return()

deliver_later(email)

deliver_later(Bamboo.Email.t()) :: Bamboo.Email.t()

deliver_now(email, opts \\ [])

deliver_now(Bamboo.Email.t(), Enum.t()) ::
  Bamboo.Email.t() | {any(), Bamboo.Email.t()}

password_reset_email(email)

password_reset_email(GitGud.Email.t()) :: Bamboo.Email.t()

Returns a password reset mail for the given email.

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email.

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. "welcome_email.text" or "welcome_email.html". Scroll to the top for more examples.

verification_email(email)

verification_email(GitGud.Email.t()) :: Bamboo.Email.t()

Returns a verification mail for the given email.