Git Gud v0.2.9 GitGud.RepoStorage

Conveniences for storing Git objects and meta objects.

Functions

cleanup(repo)

cleanup(GitGud.Repo.t()) :: {:ok, [Path.t()]} | {:error, term()}

Removes associated data for the given repo.

init(repo, bare?)

init(GitGud.Repo.t(), boolean()) :: {:ok, GitRekt.Git.repo()} | {:error, term()}

Initializes a new Git repository for the given repo.

init_param(repo)

init_param(GitGud.Repo.t()) :: term()

Returns an initialization parameter for the given repo.

push(user, repo, receive_pack)

push(GitGud.User.t(), GitGud.Repo.t(), GitRekt.WireProtocol.ReceivePack.t()) ::
  {:ok, [GitRekt.WireProtocol.ReceivePack.cmd()], [GitRekt.Git.oid()]}
  | {:error, term()}

Writes the given receive_pack objects and references to the given repo.

This function is called by GitGud.SSHServer and GitGud.SmartHTTPBackend on each push command. It is responsible for writing objects and references to the underlying Git repository.

rename(repo, old_repo)

rename(GitGud.Repo.t(), GitGud.Repo.t()) :: {:ok, Path.t()} | {:error, term()}

Renames the given repo.

workdir(repo)

workdir(GitGud.Repo.t()) :: Path.t()

Returns the absolute path to the Git workdir for the given repo.

The path is a concatenation of the Git root path, repo.owner.login and repo.name.