Git Gud v0.2.9 GitGud.ReviewQuery

Conveniences for review related queries.

Functions

comments_query(review)

comments_query(
  GitGud.CommitLineReview.t()
  | GitGud.CommitReview.t()
  | {atom(), pos_integer()}
) :: Ecto.Query.t()

Returns a query for fetching comments of a review.

commit_comment_count(repo, commit)

commit_comment_count(Repot.t(), GitRekt.GitCommit.t()) :: non_neg_integer()

Returns the number of comments for the given repo and commit.

commit_comment_count_query(repo_id, commit_oids)

commit_comment_count_query(pos_integer(), GitRekt.Git.oid()) :: Ecto.Query.t()

Returns a query for counting the number of reviews for a single commit or a list of commits.

commit_line_review(repo, commit, blob_oid, hunk, line, opts \\ [])

commit_line_review(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  GitRekt.Git.oid(),
  non_neg_integer(),
  non_neg_integer(),
  keyword()
) :: GitGud.CommitLineReview.t() | nil

Returns a commit line review for the given repo, commit, blob_oid, hunk and line.

commit_line_review_by_id(id, opts \\ [])

commit_line_review_by_id(pos_integer(), keyword()) ::
  GitGud.CommitLineReview.t() | nil

Returns a commit line review for the given id.

commit_line_review_query(id)

commit_line_review_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a single commit line review.

commit_line_review_query(repo_id, id)

commit_line_review_query(pos_integer(), pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a single commit line review.

commit_line_review_query(repo_id, commit_oid, blob_oid, hunk, line)

commit_line_review_query(
  pos_integer(),
  GitRekt.Git.oid(),
  GitRekt.Git.oid(),
  non_neg_integer(),
  non_neg_integer()
) :: Ecto.Query.t()

Returns a query for fetching a single commit line review.

commit_line_reviews(repo, commit, opts \\ [])

commit_line_reviews(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  keyword()
) :: [GitGud.CommitLineReview.t()]

Returns all commit line reviews for the given repo and commit.

commit_line_reviews_query(repo_id, commit_oid)

commit_line_reviews_query(pos_integer(), GitRekt.Git.oid()) :: Ecto.Query.t()

Returns a query for fetching all commit line reviews.

commit_review(repo, commit, opts \\ [])

commit_review(
  GitGud.Repo.t() | pos_integer(),
  GitRekt.GitCommit.t() | GitRekt.Git.oid(),
  keyword()
) :: GitGud.CommitReview.t() | nil

Returns a commit review for the given repo and commit.

commit_review_by_id(id, opts \\ [])

commit_review_by_id(pos_integer(), keyword()) :: GitGud.CommitReview.t() | nil

Returns a commit review for the given id.

commit_review_query(id)

commit_review_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a single commit review.

commit_review_query(repo_id, id)

commit_review_query(pos_integer(), pos_integer()) :: Ecto.Query.t()

commit_review_query(pos_integer(), GitRekt.Git.oid()) :: Ecto.Query.t()

Returns a query for fetching a single commit review.