Git Gud v0.2.9 GitGud.UserQuery

Conveniences for user related queries.

Functions

by_email(email, opts \\ [])

by_email(binary(), keyword()) :: GitGud.User.t() | nil

by_email([binary()], keyword()) :: [GitGud.User.t()]

Returns a user for the given email.

by_id(id, opts \\ [])

by_id(pos_integer(), keyword()) :: GitGud.User.t() | nil

by_id([pos_integer()], keyword()) :: [GitGud.User.t()]

Returns a user for the given id.

by_login(login, opts \\ [])

by_login(binary(), keyword()) :: GitGud.User.t() | nil

by_login([binary()], keyword()) :: [GitGud.User.t()]

Returns a user for the given login.

by_oauth(provider, id, opts \\ [])

by_oauth(binary(), binary(), keyword()) :: GitGud.User.t() | nil

Returns a user for the given authentication provider and id.

by_ssh_key(fingerprint, opts \\ [])

by_ssh_key(binary(), keyword()) :: GitGud.User.t() | nil

Returns a user for the given SSH key fingerprint.

search(input, opts \\ [])

search(binary(), keyword()) :: [GitGud.User.t()]

Returns a list of users matching the given input.

search_query(input)

search_query(binary()) :: Ecto.Query.t()

Returns a query for searching users.

user_oauth_query(provider, id)

user_oauth_query(binary(), binary()) :: Ecto.Query.t()

Returns a query for fetching a single user by authentication token.

user_query(id)

user_query(pos_integer()) :: Ecto.Query.t()

Returns a query for fetching a single user by id.

user_query(key, val)

user_query(atom(), term()) :: Ecto.Query.t()

Returns a query for fetching a single user by key and val.

user_ssh_key_query(fingerprint)

user_ssh_key_query(binary()) :: Ecto.Query.t()

Returns a query for fetching a single user by SSH fingerprint.

users_query(ids)

users_query([pos_integer()]) :: Ecto.Query.t()

Returns a query for fetching users by ids.

users_query(key, vals)

users_query(atom(), [binary()]) :: Ecto.Query.t()

Returns a query for fetching users by key and vals.