Git Gud v0.2.9 GitGud.Web.AuthenticationPlug

Plug providing support for multiple authentication methods.

Functions

authenticate(conn, opts)

authenticate(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with either authorization or session tokens.

authenticate_bearer_token(conn, opts)

authenticate_bearer_token(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with authorization tokens.

authenticate_session(conn, opts)

authenticate_session(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to authenticate conn with session tokens.

authenticated?(conn)

authenticated?(Plug.Conn.t()) :: boolean()

Returns true if the given conn is authenticated; otherwise returns false.

authentication_token(conn)

authentication_token(Plug.Conn.t()) :: binary() | nil

Generates an authentication token.

current_user(conn)

current_user(Plug.Conn.t()) :: GitGud.User.t() | nil

Returns the current user if conn is authenticated.

ensure_authenticated(conn, opts)

ensure_authenticated(Plug.Conn.t(), keyword()) :: Plug.Conn.t()

Plug to ensure that the request is authenticated.

If the given conn is not authenticated?/1, this prevents further plugs downstream from being invoked and returns a 401 Unauthenticated error.