Git Gud v0.2.9 GitGud.Web.Endpoint

HTTP and WebSocket endpoints.

Functions

broadcast(topic, event, msg)

Broadcasts a msg as event in the given topic.

Callback implementation for Phoenix.Endpoint.broadcast/3.

broadcast!(topic, event, msg)

Broadcasts a msg as event in the given topic.

Raises in case of failures.

Callback implementation for Phoenix.Endpoint.broadcast!/3.

broadcast_from(from, topic, event, msg)

Broadcasts a msg from the given from as event in the given topic.

Callback implementation for Phoenix.Endpoint.broadcast_from/4.

broadcast_from!(from, topic, event, msg)

Broadcasts a msg from the given from as event in the given topic.

Raises in case of failures.

Callback implementation for Phoenix.Endpoint.broadcast_from!/4.

call(conn, opts)

Callback implementation for Plug.call/2.

child_spec(opts)

Returns the child specification to start the endpoint under a supervision tree.

config(key, default \\ nil)

Returns the endpoint configuration for key

Returns default if the key does not exist.

config_change(changed, removed)

Reloads the configuration given the application environment changes.

host()

Returns the host for the given endpoint.

init(opts)

Callback implementation for Plug.init/1.

instrument(event, runtime \\ Macro.escape(%{}), fun)
(macro)

Instruments the given function.

event is the event identifier (usually an atom) that specifies which instrumenting function to call in the instrumenter modules. runtime is metadata to be associated with the event at runtime (e.g., the query being issued if the event to instrument is a DB query).

Examples

instrument :render_view, %{view: "index.html"}, fn ->
  render(conn, "index.html")
end

node_name()

An Absinthe.Subscription.Pubsub system may extend across multiple nodes connected by some mechanism. Regardless of this mechanism, all nodes should have unique names.

Absinthe invokes node_name function to get current node's name. If you are running inside erlang cluster, you can use Kernel.node/0 as a node name.

Callback implementation for Absinthe.Subscription.Pubsub.node_name/0.

path(path)

Generates the path information when routing to this endpoint.

publish_mutation(topic, mutation_result, subscribed_fields)

An Absinthe.Subscription.Pubsub system may extend across multiple nodes. Processes need only subscribe to the pubsub process that is running on their own node.

However, mutations can happen on any node in the custer and must to be broadcast to other nodes so that they can also reevaluate their GraphQL subscriptions and notify subscribers on that node.

When told of a mutation, Absinthe invokes the publish_mutation function on the node in which the mutation is processed first. The function should publish a message to the given proxy_topic, with the identity of node on which the mutation occurred included in the broadcast message.

The message broadcast should be a map that contains, at least

%{
    node: node_name,      # should be equal to [`node_name/0`](#node_name/0)
    mutation_result: …,   # from arguments
    subscribed_fields: …  # from arguments

    # other fields as needed
}

Callback implementation for Absinthe.Subscription.Pubsub.publish_mutation/3.

publish_subscription(topic, data)

After a mutation is published, and Absinthe has re-run the necessary GraphQL subscriptions to generate a new set of resolved data, it calls publish_subscription.

Your pubsub implementation should publish a message to the given topic, with the newly resolved data. The broadcast should be limited to the current node only.

Callback implementation for Absinthe.Subscription.Pubsub.publish_subscription/2.

script_name()

Generates the script name.

start_link(opts \\ [])

Starts the endpoint supervision tree.

static_integrity(path)

Generates a base64-encoded cryptographic hash (sha512) to a static file in priv/static. Meant to be used for Subresource Integrity with CDNs.

static_lookup(path)

Returns a two item tuple with the first item being the static_path and the second item being the static_integrity.

static_path(path)

Generates a route to a static file in priv/static.

static_url()

Generates the static URL without any path information.

It uses the configuration under :static_url to generate such. It falls back to :url if :static_url is not set.

struct_url()

Generates the endpoint base URL but as a URI struct.

It uses the configuration under :url to generate such. Useful for manipulating the URL data and passing it to URL helpers.

subscribe(topic)

Subscribe the current process for messages about the given topic.

Callback implementation for Absinthe.Subscription.Pubsub.subscribe/1.

subscribe(pid, topic, opts)

unsubscribe(topic)

Unsubscribes the caller from the given topic.

Callback implementation for Phoenix.Endpoint.unsubscribe/1.

url()

Generates the endpoint base URL without any path information.

It uses the configuration under :url to generate such.