Git Gud v0.2.9 GitGud.Web.ReactComponents

Functions to make rendering React components.

Functions

react_component(name)

Generates a :div containing the named React component with no props or attrs.

Returns safe html: {:safe, [60, "div", ...]}. You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent") %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component.

react_component(name, props)

Generates a :div containing the named React component with the given props.

Returns safe html: {:safe, [60, "div", ...]}. Props can be passed in as a Map or a List. You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component("MyComponent", %{language: "elixir", awesome: true}) %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component and then pass in the props specified.

react_component(name, props, attrs)

Generates a :div containing the named React component with the given props and attrs.

Returns safe html: {:safe, [60, "div", ...]}.

You can utilize this in your Phoenix views:

<%= GitGud.Web.React.react_component(
      "MyComponent",
      %{language: "elixir", awesome: true},
      class: "my-component"
    ) %>

The resulting <div> tag is formatted specifically for the included javascript helper to then turn into your named React component and then pass in the props specified.

react_component(name, props, attrs, list)