Git Gud v0.2.9 GitGud.Web.Router.Helpers

Module with named helpers generated from GitGud.Web.Router.

Functions

codebase_path(conn_or_endpoint, action, user_login, repo_name)

codebase_path(conn_or_endpoint, action, user_login, repo_name, params)

codebase_path(conn_or_endpoint, action, user_login, repo_name, oid, params)

codebase_path(conn_or_endpoint, action, user_login, repo_name, oid, id, params)

codebase_url(conn_or_endpoint, action, user_login, repo_name)

codebase_url(conn_or_endpoint, action, user_login, repo_name, params)

codebase_url(conn_or_endpoint, action, user_login, repo_name, oid, params)

codebase_url(conn_or_endpoint, action, user_login, repo_name, oid, id, params)

email_path(conn_or_endpoint, action)

email_path(conn_or_endpoint, action, params)

email_path(conn_or_endpoint, action, token, params)

email_url(conn_or_endpoint, action)

email_url(conn_or_endpoint, action, params)

email_url(conn_or_endpoint, action, token, params)

gpg_key_path(conn_or_endpoint, action)

gpg_key_path(conn_or_endpoint, action, params)

gpg_key_url(conn_or_endpoint, action)

gpg_key_url(conn_or_endpoint, action, params)

issue_label_path(conn_or_endpoint, action, user_login, repo_name)

issue_label_path(conn_or_endpoint, action, user_login, repo_name, params)

issue_label_url(conn_or_endpoint, action, user_login, repo_name)

issue_label_url(conn_or_endpoint, action, user_login, repo_name, params)

issue_path(conn_or_endpoint, action, user_login, repo_name)

issue_path(conn_or_endpoint, action, user_login, repo_name, params)

issue_path(conn_or_endpoint, action, user_login, repo_name, number, params)

issue_url(conn_or_endpoint, action, user_login, repo_name)

issue_url(conn_or_endpoint, action, user_login, repo_name, params)

issue_url(conn_or_endpoint, action, user_login, repo_name, number, params)

landing_page_path(conn_or_endpoint, action)

landing_page_path(conn_or_endpoint, action, params)

landing_page_url(conn_or_endpoint, action)

landing_page_url(conn_or_endpoint, action, params)

maintainer_path(conn_or_endpoint, action, user_login, repo_name)

maintainer_path(conn_or_endpoint, action, user_login, repo_name, params)

maintainer_url(conn_or_endpoint, action, user_login, repo_name)

maintainer_url(conn_or_endpoint, action, user_login, repo_name, params)

oauth2_path(conn_or_endpoint, action)

oauth2_path(conn_or_endpoint, action, params)

oauth2_path(conn_or_endpoint, action, provider, params)

oauth2_url(conn_or_endpoint, action)

oauth2_url(conn_or_endpoint, action, params)

oauth2_url(conn_or_endpoint, action, provider, params)

path(data, path)

Generates the path information including any necessary prefix.

repo_path(conn_or_endpoint, action)

repo_path(conn_or_endpoint, action, params)

repo_path(conn_or_endpoint, action, user_login, repo_name)

repo_path(conn_or_endpoint, action, user_login, repo_name, params)

repo_url(conn_or_endpoint, action)

repo_url(conn_or_endpoint, action, params)

repo_url(conn_or_endpoint, action, user_login, repo_name)

repo_url(conn_or_endpoint, action, user_login, repo_name, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

ssh_key_path(conn_or_endpoint, action)

ssh_key_path(conn_or_endpoint, action, params)

ssh_key_url(conn_or_endpoint, action)

ssh_key_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, token, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, token, params)