Git Gud v0.2.9 GitRekt.GitAgent

High-level API for running Git commands on a repository.

This module provides an API to manipulate Git repositories. In contrast to GitRekt.Git, it functions take and return structs such as GitRekt.GitRef, GitRekt.GitCommit, GitRekt.GitTree. Also, it allows multiple processes to manipulate a single repository simultaneously.

Example

Let's start by rewriting the example exposed in the GitRekt.Git module:

alias GitRekt.Git
alias GitRekt.GitAgent

# load repository
{:ok, repo} = Git.repository_open("/tmp/my-repo")

# fetch master branch
{:ok, branch} = GitAgent.branch(repo, "master")

# fetch commit pointed by master
{:ok, commit} = GitAgent.peel(repo, branch)

# fetch commit author & message
{:ok, author} = GitAgent.commit_author(repo, commit)
{:ok, message} = GitAgent.commit_message(repo, commit)

IO.puts "Last commit by #{author.name} <#{author.email}>:"
IO.puts message

So far, this look very similar to the original example. The real benefit of using GitRekt.GitAgent comes when multiple processes need to manipulate a single Git repository simultaneously.

Let's refactor our code for that purpose:

alias GitRekt.GitAgent

# start a dedicated process for the repository
{:ok, repo} = GitAgent.start_link("/tmp/my-repo")

count_commits = fn revision ->
  # fetch commit for given revision
  {:ok, commit, _ref} = GitAgent.revision(repo, revision)

  # walk history starting from commit
  {:ok, history} = GitAgent.history(repo, commit)

  # retrieve number of ancestors
  ancestor_cout = Enum.count(history)

  {revision, ancestor_count}
end

# simultaneously count the commit history from different revision points
~w(master my-feature-branch v0.2.8)
|> Enum.map(&Task.async(fn -> count_commits.(&1) end))
|> Enum.map(&Task.await/1)
|> Enum.each(fn {revision, count} -> IO.puts "#{revision} has #{count} commits" end)

By swapping GitRekt.Git.repository_open/1 with start_link/1, we are not working with the underlying GitRekt.Git.repo/0 anymore. Instead we use a PID to serialize function calls via message passing. This allow use to access the repository from multiple processes. In our example we start an asynchronous task for counting the number of ancestor starting from each revision and collect the result afterwards.

Note that in the above example history/2 returns a Stream.t/0 struct. We could use Stream.take/1 to retrieve the last 30 commits without having the Git agent process to enumerate the entire stream.

Repository protocol

It's also possible to implement the GitRekt.GitRepo protocol for your own data structure. For example, the GitGud.Repo schema implements this protocol and can be passed to GitRekt.GitAgent functions directly.

See attach/2 for more details on how to attach a Git agent to a struct implementing GitRekt.GitRepo.

Types

agent()

agent() :: GitRekt.Git.repo() | GitRekt.GitRepo.t() | pid()

git_object()

git_object() ::
  GitRekt.GitCommit.t()
  | GitRekt.GitBlob.t()
  | GitRekt.GitTree.t()
  | GitRekt.GitTag.t()

git_reference()

git_reference() :: GitRekt.GitRef.t()

git_revision()

git_revision() ::
  GitRekt.GitRef.t() | GitRekt.GitTag.t() | GitRekt.GitCommit.t()

Functions

attach(repo, mode \\ :inproc)

attach(GitRekt.GitRepo.t(), :inproc | :shared) ::
  {:ok, any()} | {:error, term()}

Attaches a Git agent to the given repo.

Once attached, a repo can be used to interact directly with the underlying Git repository:

{:ok, repo} = GitRekt.GitAgent.attach(repo)
{:ok, head} = GitRekt.GitAgent.head(repo)
IO.puts "current branch: #{head.name}"

Often times, it might be preferable to manipulate Git objects in a dedicated process. For example when you want to access a single repository from multiple processes simultaneously.

For such cases, you can explicitly tell to load the agent in :shared mode.

attach!(repo, mode \\ :inproc)

attach!(GitRekt.GitRepo.t(), :inproc | :shared) :: any()

Similar to attach/2, but raises an exception if an error occurs.

blob_content(agent, blob)

blob_content(agent(), GitRekt.GitBlob.t()) :: {:ok, binary()} | {:error, term()}

Returns the content of the given blob.

blob_size(agent, blob)

blob_size(agent(), GitRekt.GitBlob.t()) ::
  {:ok, non_neg_integer()} | {:error, term()}

Returns the size in byte of the given blob.

branch(agent, name)

branch(agent(), binary()) :: {:ok, git_reference()} | {:error, term()}

Returns the Git branch with the given name.

branches(agent)

branches(agent()) :: {:ok, [git_reference()]} | {:error, term()}

Returns all Git branches.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

commit_author(agent, commit)

commit_author(agent(), GitRekt.GitCommit.t()) ::
  {:ok, map()} | {:error, term()}

Returns the author of the given commit.

commit_committer(agent, commit)

commit_committer(agent(), GitRekt.GitCommit.t()) ::
  {:ok, map()} | {:error, term()}

Returns the committer of the given commit.

commit_gpg_signature(agent, commit)

commit_gpg_signature(agent(), GitRekt.GitCommit.t()) ::
  {:ok, binary()} | {:error, term()}

Returns the GPG signature of the given commit.

commit_message(agent, commit)

commit_message(agent(), GitRekt.GitCommit.t()) ::
  {:ok, binary()} | {:error, term()}

Returns the message of the given commit.

commit_parents(agent, commit)

commit_parents(agent(), GitRekt.GitCommit.t()) ::
  {:ok, [GitRekt.GitCommit.t()]} | {:error, term()}

Returns the parent of the given commit.

commit_timestamp(agent, commit)

commit_timestamp(agent(), GitRekt.GitCommit.t()) ::
  {:ok, DateTime.t()} | {:error, term()}

Returns the timestamp of the given commit.

diff(agent, obj1, obj2, opts \\ [])

diff(agent(), git_object(), git_object(), keyword()) ::
  {:ok, GitRekt.GitDiff.t()} | {:error, term()}

Returns the Git diff of obj1 and obj2.

diff_deltas(agent, diff)

diff_deltas(agent(), GitRekt.GitDiff.t()) :: {:ok, map()} | {:error, term()}

Returns the deltas of the given diff.

diff_format(agent, diff, format \\ :patch)

diff_format(agent(), GitRekt.GitDiff.t(), GitRekt.Git.diff_format()) ::
  {:ok, binary()} | {:error, term()}

Returns a binary formated representation of the given diff.

diff_stats(agent, diff)

diff_stats(agent(), GitRekt.GitDiff.t()) :: {:ok, map()} | {:error, term()}

Returns the stats of the given diff.

empty?(agent)

empty?(agent()) :: {:ok, boolean()} | {:error, term()}

Returns true if the repository is empty; otherwise returns false.

head(agent)

head(agent()) :: {:ok, git_reference()} | {:error, term()}

Returns the Git reference.

history(agent, revision, opts \\ [])

history(agent(), git_revision(), keyword()) ::
  {:ok, [GitRekt.GitCommit.t()]} | {:error, term()}

Returns the Git commit history of the given revision.

object(agent, oid)

object(agent(), GitRekt.Git.oid()) :: {:ok, git_object()} | {:error, term()}

Returns the Git object with the given oid.

peel(agent, obj, target \\ :undefined)

peel(
  agent(),
  git_reference() | git_object(),
  GitRekt.Git.obj_type() | :undefined
) :: {:ok, git_object()} | {:error, term()}

Peels the given obj until a Git object of the specified type is met.

reference(agent, name)

reference(agent(), binary()) :: {:ok, git_reference()} | {:error, term()}

Returns the Git reference with the given name.

references(agent, glob \\ :undefined)

references(agent(), binary() | :undefined) ::
  {:ok, [git_reference()]} | {:error, term()}

Returns all Git references matching the given glob.

revision(agent, spec)

revision(agent(), binary()) ::
  {:ok, GitRekt.GitCommit.t() | GitRekt.GitTag.t(), git_reference() | nil}
  | {:error, term()}

Returns the Git object matching the given spec.

start_link(arg, opts \\ [])

start_link(Path.t() | {atom(), [term()]}, keyword()) :: GenServer.on_start()

Starts a Git agent linked to the current process for the repository at the given path.

tag(agent, name)

tag(agent(), binary()) :: {:ok, GitRekt.GitTag.t()} | {:error, term()}

Returns the Git tag with the given name.

tag_author(agent, tag)

tag_author(agent(), GitRekt.GitTag.t()) :: {:ok, map()} | {:error, term()}

Returns the Git tag author of the given tag.

tag_message(agent, tag)

tag_message(agent(), GitRekt.GitTag.t()) :: {:ok, binary()} | {:error, term()}

Returns the Git tag message of the given tag.

tags(agent)

tags(agent()) :: {:ok, [GitRekt.GitTag.t()]} | {:error, term()}

Returns all Git tags.

tree(agent, revision)

tree(agent(), git_revision()) :: {:ok, GitRekt.GitTree.t()} | {:error, term()}

Returns the Git tree of the given revision.

tree_entries(agent, tree)

tree_entries(agent(), GitRekt.GitTree.t()) ::
  {:ok, [GitRekt.GitTreeEntry.t()]} | {:error, term()}

Returns the Git tree entries of the given tree.

tree_entries_by_path(agent, revision, path \\ :root, opts \\ [])

tree_entries_by_path(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  Path.t(),
  keyword()
) ::
  {:ok,
   [
     GitRekt.GitTreeEntry.t()
     | {GitRekt.GitTreeEntry.t(), GitRekt.GitCommit.t()}
   ]}
  | {:error, term()}

Returns the Git tree entries for the given revision and path.

tree_entry_by_id(agent, revision, oid)

tree_entry_by_id(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  GitRekt.Git.oid()
) :: {:ok, GitRekt.GitTreeEntry.t()} | {:error, term()}

Returns the Git tree entry for the given revision and oid.

tree_entry_by_path(agent, revision, path, opts \\ [])

tree_entry_by_path(
  agent(),
  git_revision() | GitRekt.GitTree.t(),
  Path.t(),
  keyword()
) ::
  {:ok,
   GitRekt.GitTreeEntry.t() | {GitRekt.GitTreeEntry.t(), GitRekt.GitCommit.t()}}
  | {:error, term()}

Returns the Git tree entry for the given revision and path.

tree_entry_target(agent, tree_entry)

tree_entry_target(agent(), GitRekt.GitTreeEntry.t()) ::
  {:ok, GitRekt.GitBlob.t() | GitRekt.GitTree.t()} | {:error, term()}

Returns the Git tree target of the given tree_entry.