Git Gud v0.2.9 GitRekt.Packfile

Conveniences for reading and writting Git pack files.

Types

obj()

obj() :: {GitRekt.Git.obj_type(), binary()} | {:buffer, obj_list(), obj_iter()}

obj_iter()

obj_iter() :: {non_neg_integer(), non_neg_integer(), binary()}

obj_list()

obj_list() :: [obj()]

Functions

create(repo, oids)

create(GitRekt.Git.repo(), [GitRekt.Git.oid() | {GitRekt.Git.oid(), boolean()}]) ::
  binary()

Returns a PACK file for the given oids list.

parse(pack)

parse(binary()) :: {obj_list(), binary()}

Returns a list of ODB objects and their type for the given PACK data.

parse(pack, iterator)

parse(binary(), obj_iter()) :: {obj_list(), binary()}

Same as parse/1 but starts from the given iterator.

parse_header(pack)

parse_header(binary()) :: {2, non_neg_integer()}

Returns the PACK version and the number of objects it contains.